+++ README.md
@@ -1,12 +1,3 @@
 # dums
-a terminal based dominoes game
 
-TODO : Implement
-* keep track of how many wins a player has 
-* define different win types for more points eg. milo/tell_game
-* add a function to set the score limit/boem
-* add a function to determine the order in which the game is played 
-* add a fucntion to let the user choose between teams or single player
-* visualize the dums 
-* make player hand global 
-* make modules 
+A wacky take on the original dominoes game.
