+++ main/RunGame.py
@@ -1,0 +1,15 @@
+"""
+
+This module will run the dums game
+
+"""
+
+class RunGame():
+
+    def __init__(self):
+        pass
+
+
+
+if __name__ == '__main__':
+    RunGame()
